pyproject.toml
[tool.poetry]
name = "werubin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
django-environ = "^0.9.0"
pillow = "^9.2.0"
django-crispy-forms = "^1.14.0"
djangorestframework = "^3.14.0"
django-user-agents = "^0.4.0"
django-extensions = "^3.2.1"
werkzeug = "^2.2.2"
pyopenssl = "^22.1.0"
gunicorn = "^20.1.0"
tzdata = "^2022.7"
pytz = "^2022.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
